requirements.txt
streamlit
Pillow
diffusers
torch
transformers
peft
python-dotenv